Configuration adjusted
