+++ sources/zafira-ws/src/main/java/com/qaprosoft/zafira/RootConfig.java
@@ -1,23 +1,17 @@
-package com.qaprosoft.zafira.ws;
+package com.qaprosoft.zafira;
 
 import com.qaprosoft.zafira.ws.util.dozer.NullSafeDozerBeanMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.ImportResource;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 import java.util.Arrays;
 
-//@PropertySource("classpath:environment.properties")
-@ComponentScan(basePackages = {"com.qaprosoft.zafira"})
-@EnableWebMvc
-@ImportResource(value = {"classpath:zafira-models.xml", "classpath:zafira-app-dbaccess.xml", "classpath:zafira-mng-dbaccess.xml", "classpath:zafira-services.xml"})
-//@Import(value = {ManagementPersistenceConfig.class})
+@ComponentScan
+@ImportResource(value = {"classpath:zafira-models.xml", "classpath:zafira-app-dbaccess.xml", "classpath:zafira-services.xml"})
+@PropertySource("classpath:environment.properties")
 public class RootConfig {
 
     private static final String[] DOZER_MAPPING_FILES = new String[] {
@@ -39,17 +33,15 @@ public class RootConfig {
     };
 
     @Bean
-    public PropertySourcesPlaceholderConfigurer propertyPlaceholderConfigurer() {
-        PropertySourcesPlaceholderConfigurer placeholderConfigurer = new PropertySourcesPlaceholderConfigurer();
-        placeholderConfigurer.setLocation(new ClassPathResource("environment.properties"));
-        return placeholderConfigurer;
-    }
-
-    @Bean
     public NullSafeDozerBeanMapper mapper() {
         NullSafeDozerBeanMapper dozerBeanMapper = new NullSafeDozerBeanMapper();
         dozerBeanMapper.setMappingFiles(Arrays.asList(DOZER_MAPPING_FILES));
         return dozerBeanMapper;
+    }
+
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer placeholderConfigurer() {
+        return new PropertySourcesPlaceholderConfigurer();
     }
 
 }
