+++ sources/zafira-ws/src/main/java/com/qaprosoft/zafira/WebConfig.java
@@ -1,0 +1,44 @@
+package com.qaprosoft.zafira;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+@EnableWebMvc
+@EnableAsync
+//@PropertySource("classpath:environment.properties")
+@ComponentScan(basePackages = {"com.qaprosoft.zafira"})
+public class WebConfig extends WebMvcConfigurerAdapter {
+
+    @Value("${zafira.webservice.url}")
+    private String url;
+//
+//    @Bean
+//    public PropertySourcesPlaceholderConfigurer propertyPlaceholderConfigurer() {
+//        PropertySourcesPlaceholderConfigurer placeholderConfigurer = new PropertySourcesPlaceholderConfigurer();
+//        placeholderConfigurer.setLocation(new ClassPathResource("environment.properties"));
+//        return placeholderConfigurer;
+//    }
+
+//    @Bean
+//    public static PropertySourcesPlaceholderConfigurer propertyConfigInDev() {
+//        return new PropertySourcesPlaceholderConfigurer();
+//    }
+
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("swagger-ui.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+    }
+
+}
